separate wallet from user
