+++ src/domains/mod.rs
@@ -2,3 +2,4 @@ pub mod invoices;
 pub mod lightning;
 pub mod payments;
 pub mod users;
+pub mod wallet;
