--- src/domains/users/adapters/mod.rs
@@ -1,7 +1,0 @@
-mod repository;
-mod user_balance_model;
-mod user_repository;
-
-pub use repository::UserRepository;
-pub use user_balance_model::UserBalanceModel;
-pub use user_repository::SeaOrmUserRepository;
