+++ src/domains/users/api/mod.rs
@@ -1,4 +1,1 @@
 mod auth_middleware;
-mod wallet_handler;
-
-pub use wallet_handler::WalletHandler;
