+++ src/domains/users/entities/mod.rs
@@ -1,7 +1,5 @@
 mod auth_user;
 mod permission;
-mod wallet;
 
 pub use auth_user::AuthUser;
 pub use permission::Permission;
-pub use wallet::*;
