+++ src/domains/users/services/mod.rs
@@ -1,5 +1,5 @@
 pub mod user_service;
 mod user_use_cases;
 
-pub use user_service::WalletService;
-pub use user_use_cases::*;
+pub use user_service::UserService;
+pub use user_use_cases::UserUseCases;
