+++ src/domains/wallet/adapters/mod.rs
@@ -1,0 +1,6 @@
+mod repository;
+mod user_balance_model;
+mod wallet_repository;
+
+pub use repository::WalletRepository;
+pub use wallet_repository::SeaOrmWalletRepository;
