+++ src/domains/wallet/adapters/repository.rs
@@ -1,10 +1,10 @@
 use async_trait::async_trait;
 use sea_orm::DatabaseTransaction;
 
-use crate::{application::errors::DatabaseError, domains::users::entities::UserBalance};
+use crate::{application::errors::DatabaseError, domains::wallet::entities::UserBalance};
 
 #[async_trait]
-pub trait UserRepository: Send + Sync {
+pub trait WalletRepository: Send + Sync {
     async fn get_balance(
         &self,
         txn: Option<&DatabaseTransaction>,
