+++ src/domains/wallet/api/mod.rs
@@ -1,0 +1,3 @@
+mod wallet_handler;
+
+pub use wallet_handler::WalletHandler;
