+++ src/domains/wallet/entities/mod.rs
@@ -1,0 +1,3 @@
+mod wallet;
+
+pub use wallet::*;
