+++ src/domains/wallet/entities/wallet.rs
