+++ src/domains/wallet/mod.rs
@@ -1,3 +1,4 @@
+pub mod adapters;
 pub mod api;
 pub mod entities;
 pub mod services;
