+++ src/domains/wallet/services/mod.rs
@@ -1,0 +1,5 @@
+pub mod wallet_service;
+mod wallet_use_cases;
+
+pub use wallet_service::WalletService;
+pub use wallet_use_cases::WalletUseCases;
